In the permissive bfs traversal, don't allow reverse traversal (#3717)
Just found some unintuitive traversal with the permissive BFS. For
example, when there's a graph like:

```
a -> b, c (e.g., split)
```

When traversing from just `{b}`, the normal BFS won't allow any move
since the backward traversal requires both `b` and `c`. The permissive
BFS, on the other hand, allows to visit `a` since it allows traversal
whenever there's at least one node already visited.

That's all I had in mind when I added the permissive BFS, but it turned
out that it also visits `c` as well. The first move is `b, c -> a` while
allowing the missing `c`, and the second move is `a -> b, c` since `a`
is now visited and `c` is not yet visited. This doesn't seem to make
sense since the reason `a` is visited is because we take the backward
traversal of the edge. That in turn allows the forward traversal doesn't
seem to be the right thing to do.

While I'm not aware of any particular impact due to this behavior, this
PR prevents such traversal by checking if any of Val nodes is already
marked as a previous node of an Expr node.