Print scheduled fusion instead of kernel for NVFUSER_DUMP=fusion_ir (#3769)
PR #3468 changed to using `CompiledKernel` and in the shuffle, we
switched from using the incoming fusion before compilation to the
lowered `kir::Kernel`. This PR just moves the printing to occur just
before lowering, inside the constructor for `CompiledKernel`. I believe
this is enough to restore the previous behavior.

Fixes #3765