Fix stmatrix scheduling for persistent GEMM (#3791)
## Problem
The current stmatrix scheduling assumes all iterDomains are parallelized
to the left of the mma output allocation domain.
Therefore, it moves the stmatrix serial iterDomain to the 0th position.
This is incompatible with persistent gemm kernels, which have a grid
strided serial iterDomain.

## Solution
This PR fixes this moving the stmatrix serial iterDomain back one
position and using the 3rd from the end for-loop during index
generation. The 3rd from the end for-loop is the 0th position from the
mma output allocation domain.