Naive register <-> tmem load/store support (#3786)
Extracted from https://github.com/NVIDIA/Fuser/pull/3755 to make code
review easy.

This PR adds a new unit test `TMemTest.GmemRegTMemRegGmemCopy` that
schedules a copy kernel gmem -> register -> tmem -> register -> gmem,
and update our system with the minimum required changes to make this
test pass.

The purpose of this PR is not to provide a good implementation of TMem
support, but just to provide the absolute minimal requirement for us to
start. Limitations are:
1. The index is hard coded zero, so this PR is not touching the
interesting topic of "how to schedule TMem tensor?"
2. The TMem is used without allocation. Using a memory that is not
allocated is clearly a wrong way to program, but as described in the
code comment, if a fusion only has one TMem TensorView, it is guaranteed
to work.

Generated code:
```CUDA
__global__ void nvfuser_none_f0_c0_r0_g0(Tensor<float, 1, 1> T0, Tensor<float, 1, 1> T4) {
  nvfuser_index_t i0;
  i0 = ((nvfuser_index_t)threadIdx.x) + (32 * ((nvfuser_index_t)blockIdx.x));
  bool b1;
  b1 = i0 < T0.logical_size[0LL];
  Array<float, 1, 1> T1;
  T1[0] = 0;
  if (b1) {
    T1[0]
       = T0[((T0.alloc_stride[0LL] * ((nvfuser_index_t)threadIdx.x)) + ((32 * T0.alloc_stride[0LL]) * ((nvfuser_index_t)blockIdx.x)))];
  }
  asm volatile(
    "tcgen05.st.sync.aligned.32x32b.x1.b32 [%0], {%1};\n"
    :
    :"r"(0U),
     "f"((*reinterpret_cast<Array<float, 1, 1>*>(&T1[0]))[0])
  );
  asm volatile("tcgen05.wait::st.sync.aligned;\n");
  Array<float, 1, 1> T3;
  asm(
    "tcgen05.ld.sync.aligned.32x32b.x1.b32 {%0}, [%1];\n"
    :"=f"((*reinterpret_cast<Array<float, 1, 1>*>(&T3[0]))[0])
    :"r"(0U)
  );
  asm volatile("tcgen05.wait::ld.sync.aligned;\n");
  if (b1) {
    T4[i0]
       = T3[0];
  }
}
```