Add options for persistent kernels/warp specialization in MatmulParams (#3792)
This adds the ability to request warp specialization and persistent
kernels in MatmulParams, including in the python frontend. Note that the
functionality of persistent CTA scheduling will be added in a follow-up
PR. For now, we throw an error when persistent kernels are requested.

This PR also parametrizes the three MLP Benchmark tests used in our
recent perf dive for convenience.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>