`tcgen05.alloc` TMem usage (#3795)
Stacked on https://github.com/NVIDIA/Fuser/pull/3786, extracted from
https://github.com/NVIDIA/Fuser/pull/3755 to make code review easy.

This PR adds the missing `tcgen05.alloc` and
`tcgen05.relinquish_alloc_permit` in the kernel. Like in
https://github.com/NVIDIA/Fuser/pull/3786, we are still limited that
each fusion only have one TMem TensorView. But this time, because we are
relinquishing the right to allocate, CTAs are no longer serialized.

Because the `tcgen05.alloc` writes the allocated TMem address to smem,
there are some tricks here to make things work:
1. We need an "address tensor" to store the address. A `Val` can not be
placed on shared memory. The index for this address will always be zero.
2. We need to pay special attention to make sure that our stack based
smem allocator can correctly allocate smem for the address tensor. For
example, we could not just insert `kir::Asm` for `tcgen05.alloc` in the
allocation pass. We need a new expression `kir::AllocTMem` whose output
is the address tensor, so that our first-write and last-read analysis
works.

Generated kernel:
```CUDA
__global__ void nvfuser_none_f0_c0_r0_g0(Tensor<float, 1, 1> T0, Tensor<float, 1, 1> T4) {
  alignas(16) extern __shared__ char array[];
  const unsigned smem_offset = 0;
  nvfuser_index_t i0;
  i0 = ((nvfuser_index_t)threadIdx.x) + (32 * ((nvfuser_index_t)blockIdx.x));
  bool b1;
  b1 = i0 < T0.logical_size[0LL];
  uint32_t* T5 = reinterpret_cast<uint32_t*>(array + smem_offset + 0);
  asm volatile("tcgen05.alloc.cta_group::1.sync.aligned.shared::cta.b32 [%0], %1;\n"::"r"((uint32_t)(toSmem(T5))), "n"(32U));
  asm volatile("tcgen05.relinquish_alloc_permit.cta_group::1.sync.aligned;\n");
  __syncthreads();
  Array<float, 1, 1> T1;
  T1[0] = 0;
  if (b1) {
    T1[0]
       = T0[((T0.alloc_stride[0LL] * ((nvfuser_index_t)threadIdx.x)) + ((32 * T0.alloc_stride[0LL]) * ((nvfuser_index_t)blockIdx.x)))];
  }
  TMemTensor T2(T5[0]);
  asm volatile(
    "tcgen05.st.sync.aligned.32x32b.x1.b32 [%0], {%1};\n"
    :
    :"r"((uint32_t)(T2 + Array<uint16_t, 2, 1>{0, 0})),
     "f"((*reinterpret_cast<Array<float, 1, 1>*>(&T1[0]))[0])
  );
  asm volatile("tcgen05.wait::st.sync.aligned;\n");
  Array<float, 1, 1> T3;
  asm(
    "tcgen05.ld.sync.aligned.32x32b.x1.b32 {%0}, [%1];\n"
    :"=f"((*reinterpret_cast<Array<float, 1, 1>*>(&T3[0]))[0])
    :"r"((uint32_t)(T2 + Array<uint16_t, 2, 1>{0, 0}))
  );
  asm volatile("tcgen05.wait::ld.sync.aligned;\n");
  if (b1) {
    T4[i0]
       = T3[0];
  }
}
```