Fix max logical domain propagation with resize (#3840)
Fixes #3801.

Please see the repro of #3801:
https://github.com/NVIDIA/Fuser/issues/3801#issuecomment-2638101547

Here's how the fusion looks like:


![repro3801_fusion](https://github.com/user-attachments/assets/46874cc7-a58b-47c6-9279-c5473924c411)

The issue with the repro is that T10 is not inlined at all. The sync
error is a side effect of the lack of inling.

The reason T10 isn't inlined at all is because `getPositionsMappedTo` of
inlining.cpp returns 0 for T10, which is because the spanning tree
traversal comes from the reference tensor of T11 through T1, which means
`FindMappedPositions` is computed through the Resize exprs for the slice
ops. Since inlining is not allowed across resized IDs, the resulting
position is propagated as 0.

However, in this case, we can actually inline T10 safely. The inlining
position returned by the analysis is not correct. In fact, there's a
traversal path that doesn't go through the resize ops, i.e., from T11
through T3 and T13 then T10. If this path were taken, the analysis would
have allowed inlining of T10.

To fix the issue, for inlining, we should prefer paths that don't
involve Resize exprs. I thought we should always do this, but I realized
that for the sake of transform propagation we don't want to deprioritize
Resize, since we are currently propagating slice and merge ops across
resized IDs rather permissively. I consider this behavior is a technical
debt and we should take care of resize ops more explicitly as is done by
the resize scheduler. For the time being, to keep the current behavior
as is, `MaxLogicalDomainInfoSpanningTree` should not propagate through
Resize exprs for inlining only.

To change the behavior of `MaxLogicalDomainInfoSpanningTree`, I added a
flag, `propagate_through_resize_`, to
`MaxLogicalDomainInfoSpanningTree`. I don't like the idea of adding a
flag since it feels like an ad-hoc fix, but I couldn't find any other
reasonable alternative.