Multidimensional device mesh (#3821)
1. Extends DeviceMesh to support multidimensional meshes. Representation
is a flat vector of devices and a vector of shapes. Since we only have
plans to support DIDx, DIDy, and DIDz this naturally restricts the max
dimensions to 3D.
2. Key functionality is in `getSlice(DeviceIdxType, ParallelType)` which
gives a set of devices that a tensor is sharded over and communicates
with given the (device) parallel type which translates into a dimension
into the mesh.
3. Trivially modifies communication lowering to analyze a slice instead
of the entire mesh for collectives that use the same mesh (AllReduce,
ReduceScatter, AllGather). Collectives without this property assert that
it is operating over a 1D mesh.
4. Adds ParallelType's DIDy, DIDz. These are not used anywhere except to
index into 2D and 3D device mesh tests.