Loop promotion for cyclic graphs (#3747)
Closes #3072 

ValGraphStmtSort is currently used in two places:

- `computeCoveredGroups`
- `propagatePromotionsInIELGraph`

ValGraphStmtSort can ignore cycles due to the almost-exact mapping, but
that was it. This PR extended the visitor to support cycles induced by
reshapes and resizes. It simply ignores cyclic dependencies, so the user
of the visitor needs to make sure that's sufficient for respective
analyses. For the above two cases, that should be valid.

Just ignoring cycles during traversal is not enough as there may be no
terminating inputs. In order to circumvent the issue, I added
`getInputGroupsOfExactGraph` and `getInputGroupsOfIELGraph`. These two
functions get val groups that should be used as starting groups for the
respective graphs. These input groups supplement the terminating inputs
in ValGraphStmtSort.