Add optional lowering to Host IR in FusionKernelRuntime (#3835)
This PR adds milestone 2 from Jingyue's document [Initial Integration of
Host IR to
FusionExecutorCache](https://docs.google.com/document/d/1QrRmN27XsVjZu7QrZWJJyRENO50878LC3MvlQY1cRYA/edit?usp=sharing)

It adds an optional codepath in FusionKernelRuntime (FKR) to lower a
fusion to Host IR. When enabled, FKR will create a HostIrEvaluator,
which has a container that all of the fusion's expressions are pushed
to. Expressions lowered to a kernel are launched with the LaunchKernel
HostIr.