Allow 64-bit indexing for TMA instructions, with validation (#3850)
This extends #3599 by also computing the minimal dtype required by the
expressions in the lowered kernel. Like in #3599, we cast from
`nvfuser_index_t` to `int32_t` when passing coords to the TMA
expression. However, _unlike_ #3599 we actually verify that this is safe
to do by checking the bounds of the inputs to those casts. This
validation is done for every new set of inputs, when we initialize a new
`ExecutorEntry`.

Most of the code in this PR is validation code that resides in
`interval_analysis.{cpp,h}` and `test_interval_analysis.cpp` which
implements `executor_utils::validateIndexCasts` which itself is called
in `executor.cpp`. If we are willing to forego validation, we could
instead just merge the changes to `matmul_utils.cpp` that remove the
rejection code and `index_compute.cpp` which inserts the int32 cast for
TMA coords.

Note that this PR does not change the index type of the kernel based on
the bounds analysis, so currently large matmuls will still use 64-bit
indexing even when that's not necessary. Setting the index type properly
can be done in a future PR using this interval analysis machinery.

Fixes #3601

---------

Co-authored-by: Naoya Maruyama <naoyam@users.noreply.github.com>