Implement persistent matmul scheduling (#3812)
Stacked on #3642 

This is a followup to #3792 that implements persistent scheduling.

There is a current limitation that affects both persistent scheduling
and "grid swizzling": if `MatmulOp` or `LinearOp` are present in the
fusion, we will hit inlining errors. This is because in that case we
have a non-trivial `AxisMapping` on the `MmaOp`. The missing input
dimensions are not tracked through the scheduling transforms (merges and
splits) required for either grid swizzling or persistent scheduling.
Because of this, I introduced three new parametrized tests matching the
original `MLPBenchmarkTest`s but with `_BroadcastInputs` suffix. These
tests use `fusedMultiplySum` instead of `linear`. The `persistent`
variant of the non `BroadcastInputs` tests are skipped until we fix the
inlining issue.

I currently observe a correctness issue in the
`MLPBenchmarkTest.FwdEpilogueFusion_BroadcastInputs` test regardless of
parametrization. This means that we are getting incorrect results even
for data parallel scheduling. I confirmed this test also fails on main.
I currently skip this test with a warning mesage.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>