Resize scheduler vectorization WAR (#3955)
This is a WAR for an issue with the vectorization by the resize
scheduler (unrelated to #3640).

#3693 introduced a reordering optimization for the resize scheduler that
attempted to minimize strides in read accesses of fusion inputs by
canceling reshapes. It turned out it can result in conflicts with
vectorization. The scheduler uses the fusion input as the reference of
the vectorization analysis, assuming any reshape is canceled, which is
not always the case.

So, in this PR, the vectorization analysis is changed to use the
reference output. However, that isn't enough since when a resize is
indeed canceled, the analysis should actually be done using the
pre-reshape shape.

To workaround that, this PR also adds a flag to disable canceling
reshapes that use innermost logical IDs. This should make sure it's
always valid to use the fusion output as the reference of the
vectorization analysis.

This is an ad-hoc WAR but should be good enough for the RoPE cases. The
real problem is a bit inter-twinned here, and I'm not attempting to
address it completely in this PR.