Tensor memory allocation analysis (#3922)
This PR updates the logic in the tensor memory lowering analysis to
remove the hardcoded "32 columns". We are now doing a real analysis on
the allocation domain, and compute the number of columns based on the
schedule. This PR also adds validations of the allocations.

As a result, we can enables two tests: `TooManyLanes`, `TooManyCols`.