Construct `FusionExecutorCache`s with a ready-to-run fusion (#3969)
As requested by myself at
https://github.com/NVIDIA/Fuser/pull/3923#discussion_r1964932407

This PR tries to fix the Python frontend to construct
`FusionExecutorCache`s with a ready-to-run fusion. Previously, the code
constructs FusionExecutorCache with an empty fusion and populates it
later. This is not how we normally use and test FusionExecutorCache and
is indeed problematic in some situations. For example,
`FusionExecutorCache::exact_map_` was always empty because the
constructor thought it was going to run an empty fusion. See one of the
tests fixed by this PR.

cc @samnordmann