Prevent self mapping in the AlmostExact graph (#3926)
Fixes #3919 

The root cause of #3919 is there's a tensor that has a self mapping loop
domain. Specifically:

```
// [1, 1]
tv0: [b1, b2]
// [3,3]
tv1 = pad(tv0, {1, 1, 1, 1});

tv0->merge(0, 1);
```

In this case, because the tv0 merge just uses broadcast IDs, all of the
two input IDs and the output ID are mapped in the AlmostExact graph (see
[here](https://github.com/NVIDIA/Fuser/blob/main/csrc/id_model/id_model.cpp#L435-L440)).
That is also the case with
[ComputeAtMap](https://github.com/NVIDIA/Fuser/blob/main/csrc/compute_at_map.cpp#L738-L748).
That means `b1` and `b2` are mapped, which in turn means the two logical
IDs of tv1 are also mapped since they use the same resize op.

While the mapping of `b1` and `b2` may not have any actual effect as
they are just broadcast IDs, the mapping of the logical IDs of tv1 is
problematic since they are concrete IDs. In the case of #3919, that
manifests as the error in the contiguity analysis since it assumes no
self mapping.

The fix in this PR is simply avoiding mapping `b1` and `b2`. More
specifically, I extended `ValGraph` to exclude certain Vals from
mapping. In the case of the AlmostExact graph, even when two IDs are
trivially mapped, we no longer map them if they can result in self
mapping. It may not sound ideal as they do have the same extent, but
until we address the self mapping limitation, I think this is a
reasonable workaround.

Note: I am not 100% confident if this might have any negative side
effect. That is, the AlmostExact graph now lacks some mappings that do
have the same extent. However, self mapping would certainly break
TensorIndexer, so I think this is still reasonable.