Remove trivial G->G broadcasts at lowering (#3968)
In this PR I am introducing new behavior in lowering which skips
`BroadcastOp`, `SqueezeOp`, or `LoadStoreOp` when
1. both the inputs and outputs lie in global memory
2. both the inputs and outputs are contiguous
3. the output is not a Fusion output
4. their allocation domains "match": i.e. when all non-broadcast
non-reduction IDs are exact mapped and in the same stride order.

The intention behind these conditions is to isolate ops for which it is
safe to avoid moving data into a temporary global buffer when no
computation is performed and all elements are in the same relative
linear positions in that intermediate buffer. Instead, we will use the
producer buffer as an alias to avoid a needless G->G copy.
**Importantly, this allows us to schedule G->G operations in order to
_view_ Fusion inputs as broadcast/squeezed/permuted logically** without
needing to codegen new operations and incur avoidable data movements.

When these operations are skipped, no `Allocate` nodes are created
meaning we do not actually allocate the intermediate global tensor, and
we do not generate any code for initializing or copying data into those
global tensors. We generate a linear index for the _unaliased_ tensor
whenever it is used, but then we replace the `view` of the `TensorIndex`
to point to the alias target.

This PR is motivated by #1628 and #3671. It is an alternative to #3366
and #3372, and would allow us to roll back #3391, #3414, #3416. In a
follow-up, I will create an alternative to #3440 which translates
`MatmulPattern` on Hopper, which I believe will let us fix #3671 and
enable persistent matmul with `MatmulOp` and `LinearOp`.