DID loop split on reshaped IDs (#3875)
This PR updates `propagateReshapeTransform` to support DID loop split.

When the loop split is on the iterdomains being reshaped, the logical
reshaped iterdomain is no longer present in the loop domain since it is
split. In this case, we check if there is a sharded loop ID and compare
the logical reshape iterdomain to the producer of this DID split.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>