Relax BFS target constraint in getInnerMmaLoopGroup (#4012)
In the indexing lowering pass, getInnerMmaLoopGroup is used to determine
the inner and outer MMA strides, which is used to create the wgmma
descriptor. The purpose of that function is to obtain the ValGroup of
the consumer loop ID corresponding to the innermost allocation domain of
the producer. So we do need to be able to traverse from consumer loop to
the innermost allocation group in the ValGraph, but we do not care about
visiting any other groups.

See https://github.com/NVIDIA/Fuser/issues/3962#issuecomment-2695415388
for an example where we cannot currently reach some of the outer
allocation dimensions if grid swizzling is used.

This PR just sets `require_all_to_visited` to false when performing the
BFS. A more involved fix might update
`ValGraphBFS::getExprGroupsBetween` to accept a vector of required
groups to be visited. Even better would be to understand fully why we
are unable to visit the grid swizzled allocation domains and address
that instead.

Fixes #3962