Update Hopper default matmul heuristic (#4022)
On main, our default Hopper heuristic gets poor perf:

![scurve_oldheur305](https://github.com/user-attachments/assets/f92d7efb-75fa-451f-a707-21be377aab81)

This PR does some optimization to determine better tile sizes and load
stages:
1) We always use instruction tile M/N equal to warp tile M/N, but warp
tile K can vary from 16 to 256
2) We always use 1 or 2 math groups. When 1 math group is used the CTA
tile matches the warp tile. When 2 math groups are used, the CTA tile is
double the warp tile in either the M or N dimension.
3) We choose tile sizes by estimating the total amount of data
transferred in the fusion, ignoring caching altogether. The
configuration that minimizes bytes transferred is selected. **NOTE:
currently we require all M, N, and K sizes to be multiples of 64. Other
configurations are valid, but I got poor perf when enabling them.**
4) After choosing tile size, we maximize load stages, staying within the
smem limits of the device.

As of d6d359be:

![scurve_newheur0306_largermacro_allmul64](https://github.com/user-attachments/assets/86033fe8-8f9e-423e-80b9-388e52c76cdc)

### Future improvements

In a future PR I would like to introduce a naive model of L2. This will
be particularly useful for selecting the grid swizzle pattern.