[python] return output tensors and output shardings separately (#4035)
[Unpacking](https://github.com/NVIDIA/Fuser/blob/22d4f9702ea4a13028776b63e5ca39fcf8d3edc1/nvfuser/__init__.py#L328)
in Python is apparently slow.

This PR changes FusionDefinition.execute to return the output tensors
and the output shardings separately. This way, single-GPU programs don't
pay the overhead of packing/unpacking shardings.

As a side benefit, this allows FusionDefinition::execute to return
KernelArgumentHolder, making it possible for FusionDefinition.execute to
return scalars.