Tensor memory 32x32b data path pattern matching (#4015)
This PR adds pattern matching for the data path `32x32b`, where each
warp access a subpartition of the tensor memory. In short, the algorithm
for this pattern matching is to use:
```C++
Val* stride = lower_utils::proveLinearAndGetStride(id_graph, warp, lane_allocation_domain);
```
If `stride` is `nullptr`, this means we failed to prove the warp is
linear in the lane allocation domain. For this case, we can not say that
the data access pattern is `32x32b`. If the `stride` is not `nullptr`,
then it must be `1` according to hardware specification for `32x32b`. If
not `1`, then this is not a `32x32b`.

A tricky thing to implement the above algorithm is: what is `warp`?
Depending on the size of parallel dimensions, the answer can be
different: `warp` may be just `TIDx`, or `TIDx % 32`, or `(TIDy * TIDx)
% 32`, etc. This PR includes an analysis that handles all these cases.
See the PR for detail.