Globally use `TensorIndexer` if TMem is used (#4030)
Currently, the `TMemTutorialR.Complicated1` and
`TMemTutorialR.Complicated2` are skipped because silent wrong results
are computed. The reason for the silent wrong results was not because
the lowering of TMem was incorrect, but the opposite: TMem was lowered
correctly, however, our legacy indexer was unable to handle schedule
generated by `fancyTransformations`.

In this PR, I just draw a line saying that if you want to use TMem, you
have to use `TensorIndexer`. This requirement may sounds too strict, but
we are retiring the legacy indexer anyway, and I don't mind making it
die faster by making up an arbitrary rule.