Fix C++23 backport of `zip` and `enumerate` (#4068)
The current implementation can not be composed with `reverse`, like
`zip(v, w) | std::views::reverse`, which is different from C++23's
behavior. This PR fixes that issue by add the missing `operator--` when
the input views are bidirectional. I also added `enumerate_view` and
uses it to implement `enumerate`. This implementation is closer to how
C++23 implements `enumerate`.