Safe resize vectorization (#3906)
Fixes #3640.

The current issue with resize vectorization is that because of the
spanning tree based traversal, not all paths are taken into
consideration when determining vectorization factors. This PR addresses
the issue by finding all `Resize` ops that have dependencies with
vectorized IDs. To do so, added a new graph traversal class,
`FindAllExprs`, based on `BFS`. Note that `BFS` only finds shortest
paths, so not suitable when, for example, `Resize` appears in a cycle,
which is the case with RoPE.

I was also thinking about overhauling the vectorization analysis to be
more ID traversal basis, but we would still need to look at each
tensor's allocation domain, so I don't think we can completely move away
from tensor op traversals.