Update SDPA flash attention API (#4065)
https://github.com/pytorch/pytorch/pull/146372 changes the flash
attention API.

```
// [Note] BC breaking change to flash seed/offset
// Previously: Used separate tensors for philox_seed and philox_offset, sometimes on CPU, sometimes on CUDA
// FlashAttention change: Now uses a single uint64_t[2] tensor on device containing both seed and offset
// Implementation: Renamed "seed" → "rng_state" (contains both seed+offset) and "offset" → "_unused"
```

~In nvfuser API, `SdpaFwdOp` now returns a `rng_state` and `SdpaBwdOp`
expects a `rng_state`. `_unused` is ignored.~

In nvfuser API, based on torch version, `philox_seed` is now
`uint64_t[2]` and philox offset is a empty `uint64_t` tensor on device.
I chose to keep the same semantics as PyTorch since it avoids
packing/unpacking PyTorch outputs when testing if I were to split the
`rng_state` output of PyTorch into `philox_seed` and `philox_offset`.
The latter approach would keep the dimensions of both tensors the same,
but since we're changing `device` and `dtype`, they will need to be
created distinctly for the two cases. This also allows us to switch to
the single-parameter version in the future more easily if desired.

---------

Co-authored-by: Jingyue Wu <wujingyue@gmail.com>