WAR for false ComputeAtLogicalDomainMap results (#4083)
Fixes #4020.

`ComputeAtLogicalDomainMap` falsely flags some tensors as having
unmappable iter domains, resulting marking those tensors as persistent,
although they should not need to be persistent. In the case of #4020,
this results in failing to find a resolution point of a false persistent
buffer because there's no resolution.

See `NVFuserTest.FalsePersistentBuffer` for a simplified example. With
the current main, the test fails at the call to `persistentBuffers`
since it can't find a resolution point of `tv0`.

I think the root case of this problem is how reshape is handled in
`ComputeAtLogicalDomainMap`, specifically, for each reshape output
tensor, it tries to aggregate all mappings propagated from consumer
tensors to the logical IDs of the tensor, and then the mapping info is
aggregated to the corresponding root IDs. This doesn't seem to make
sense, and I don't remember why (likely me) did this.
`ComputeAtLogialDomainMap` was originally added before reshape support
was added, so it was not originally designed for anything related
reshape. In any case, this class is due for a significant update.

In the meantime, this is a quick WAR to unblock the original issue.
Whenever a resolution point is not found, we should not return an error
but gracefully fall back to other schedulers.