Recompute non-cacheable unmappable tensors (#4078)
This PR is meant to address the issue of caching persistent buffers as
found in #4074 and #4052.

The problem is caused by a persistent buffer when it's broadcast and
concretized to a reduction ID. Since the broadcast ID is inlined into
the concrete, reduction ID, which by definition has no mapped ID in the
persistent buffer, the persistent buffer may be inconsistently
parallelized, thus resulting in the error at the sync analysis.

There are a couple of potential approaches to address the issue. For
example, we could artificially expand the persistent buffer with an ID
that is exactly mapped with the reduction ID to which the broadcast ID
is concretized. This should be possible through
`TensorDomain::setLoopDomain`, but this way of scheduling is not yet
extensively exercised.

Another approach is to give up caching such tensors, which is what this
PR does. This isn't ideal either, but may not matter much in practice.
We haven't encountered this issue until #4052, which indicates this is a
pretty rare pattern, so I think that to be conservative is likely just
fine. Moreover, in the case of #4052, this would only disable caching of
`T0` and `T1` should still be cached. Since `T0` is much smaller, it
might not matter much in practice.

TODOs:

- [x] Don't do this when cached on shared memory, since in that case we
can resolve the consistency by a sync
  - Not in this PR. #4090 
- [x] Don't project to inputs that may not be cacheable. This is
necessary when the issue is triggered by the input projection like the
original repro of #4052.
- [x] Support reshape (BroadcastSyncReshape is currently failing)
  - Not in this PR. #4097

---------

Co-authored-by: Liqiang Lu <liqiangxl@gmail.com>
Co-authored-by: Liqiang Lu <116412316+liqiangxl@users.noreply.github.com>