Reduce the number of iterations
Each test takes >50 seconds on CI currently. While I understand more
iterations are needed to get a stable reading on time, CI doesn't
read time so doing so by default is a waste.

Consider adding a flag to control the number of iterations, or create a
separate benchmark with the fusion definition extracted.