TMem check the stride of outer dims (#4070)
Validating that the TMem ld/st is scheduled as a warp-collective is
fundamentally the same problem as validating the vectorization of a
slice op with vectorization factor 32. This validation requires us to
check the inner dim size, as well as outer dim's strides. Currently, we
are only checking the inner-dim's size. This PR adds the missing stride
check.

See discussion:
https://github.com/NVIDIA/Fuser/pull/4015#discussion_r1991999815