Enable the new indexer for gather-like indirect indexing (Part 1) (#4108)
This PR enables indexing of ops like `index_select` and
`take_along_axis` with `TensorIndexer`. This first PR adds minimal logic
with the following restrictions:

- No support for `gather`. Both `gather` and `take_along_axis` are
represented as `GatherOp` with the only difference of the
[exactness](https://github.com/NVIDIA/Fuser/blob/main/csrc/ir/internal_nodes.h#L136-L138).
Only the exact case is supported.
- No support for indexing intermediate tensors. Some more refactoring of
allocation domains is needed.
- No support for scatter. I actually wonder if we should just remove the
current implementation as there's probably not much we could reuse.