Introduce use_stmatrix parameter to MatmulParams (#4100)
This just disables stmatrix when the problem size is not a multiple of
M=16 N=16. Note that we still use TMA for the epilogue when possible.

Fixes #3963. This is actually a work-around but might be the final fix
unless we have important perf to gain from a proper fix for
non-divisible sizes.

---------

Co-authored-by: Ryan Spring <rspring@nvidia.com>