Fix indexing of take_along_axis (#4128)
Previously, in TensorIndexer, indexing indirectly accessed tensors in
ops like take_along_axis wasn't working when the tensor is not entirely
allocated (in global memory). It was due to the allocation domain not
properly set. Specifically, since the indirectly accessed logical ID is
indexed, it must be included in its allocation domain. The legacy
indexer somehow works, but the new indexer resulted in a failure with
the index path traversal. This PR fixes the allocation domain.

Note that `gather` is not yet supported in TensorIndexer when it is not
exact.