Fix liveness of TMem address smem TV (#4142)
To use TMem, we must have a smem tensor for storing the address of TMem,
and uses `smem_tv[0] = tcgen05::alloc()` to store the allocated TMem
address. The liveness of this smem tv should be from the
`tcgen05::alloc` to the `tcgen05::dealloc`. However, we are currently
unable to get that due to `tcgen05::dealloc(smem_tv[0])` is not a "TV
Op". Due to this bug, the smem address tensor is allocated overlapping
with other smem tensors. This PR fixes this bug by handling both TV Op
or non-TV op correctly.